refactor(desktop): extract the link-paste trailing-space extension
useRichTextEditor.ts sat at 999 lines on main, so this branch's six added
lines tripped the desktop file-size ratchet. Move the self-contained
LinkPasteTrailingSpace extension into a sibling module, matching the
existing extension-per-file layout, rather than raising the ceiling.

Pure move: the extension body is unchanged.
